+++ tests/unit/structures/test_agent.py
@@ -286,10 +286,6 @@ class TestAgent:
         assert agent.tasks[0].prompt_driver.stream is False
         assert agent.tasks[0].prompt_driver is not prompt_driver
 
-    def test_validate_stream(self):
-        with pytest.warns(UserWarning, match="`Agent.stream` is set, but `Agent.prompt_driver` was provided."):
-            Agent(stream=True, prompt_driver=MockPromptDriver())
-
     def test_validate_prompt_driver(self):
         with pytest.warns(UserWarning, match="`Agent.prompt_driver` is set, but `Agent.stream` was provided."):
             Agent(stream=True, prompt_driver=MockPromptDriver())
@@ -298,17 +294,20 @@ class TestAgent:
         with pytest.warns(UserWarning, match="`Agent.tasks` is set, but `Agent.prompt_driver` was provided."):
             Agent(prompt_driver=MockPromptDriver(), tasks=[PromptTask()])
 
-    def test_sugar_fields(self):
+    def test_field_hierarchy(self):
+        # Test that stream on its own propagates to the task.
         agent = Agent(stream=True)
 
         assert isinstance(agent.tasks[0], PromptTask)
         assert agent.tasks[0].prompt_driver.stream is True
 
+        # Test that stream does not propagate to the prompt driver if explicitly provided
         agent = Agent(stream=True, prompt_driver=MockPromptDriver())
 
         assert isinstance(agent.tasks[0], PromptTask)
         assert agent.tasks[0].prompt_driver.stream is False
 
+        # Test that neither stream nor prompt driver propagate to the task if explicitly provided
         agent = Agent(
             stream=False,
             prompt_driver=MockPromptDriver(stream=False),